third_party/WebKit/LayoutTests/http/tests/misc/client-hints-picture.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
    var counter = 1;
    var t = async_test("Verify that write width hint is sent when picture is involved.");
    window.addEventListener("message", t.step_func(function (message) {
        if (message.data.indexOf("fail") != -1) {
            assert_unreached("The test failed on " + message.data);
            t.done();
            return;
        }
        if (counter == 1) {
            assert_equals(message.data, "resize");
            ++counter;
            document.getElementById("frame").style.width = "600px";
        } else {
            assert_equals(message.data, "success");
            t.done();
        }
    }));
</script>
<iframe src="resources/iframe-ch-with-picture.php" onerror="console.log('failed to load iframe');t.done()" id=frame style="width: 800px"></iframe>
